lib/run: pulled out of update
Promise based command line
