+++ lib/update.js
@@ -2,20 +2,11 @@ var q = require('q'),
 app = require('flatiron').app,
 path = require('path'),
 minimatch = require('minimatch'),
-child_process = require('child_process');
-
-function run(command) {
-	return q.nfcall(
-		child_process.exec,
-		command,
-		{ cwd: app.config.get('cwd') }
-	);
-}
+run = require('./run');
 
 function fetch() {
 	return run('git status --porcelain')
-	.then(function(stdout) {
-		app.log.verbose.apply(app.log, stdout);
+	.then(function() {
 		return run('git fetch');
 	});
 }
@@ -29,7 +20,6 @@ function diff(ref) {
 	var isSource = minimatch.filter(path.join(app.config.get('source'), '**'));
 	return run('git diff --name-only ..' + ref)
 	.then(function(stdout) {
-		app.log.verbose.apply(app.log, stdout);
 		// Affects the application
 		return stdout && !stdout.every(isSource);
 	});
